// apps/webapp/src/interfaces/CreateContractBody.ts
interface CreateContractBody {
  studentsNanme: string;
  schoolName: string;
  dateBirthStudent: string;
  curp: string;
  studentPhone: string;
  scholarship: string;
  startDateService: string;
  modality: string;
  annualRegistration: string;
  monthlyPayments: [];
  courseId: string;
}

export default CreateContractBody;